refactor: extract RecordingSessionOutcomes.swift (#650 slice B)
First slice of #650 (RecordingSessionController.swift decomposition). Extracts
the 3 outcome enums (RecordingSessionStartOutcome, RecordingSessionStopReadiness,
RecordingSessionCancelOutcome) from RecordingSessionController.swift into a
new RecordingSessionOutcomes.swift.

Byte-identical move; SHA-256 33986becd1b38a87922f06443726f36fcefaa2c310b2f1f197a88ec62430339c
verified against origin/main:RecordingSessionController.swift lines 408-427.

Reduces RecordingSessionController.swift from 427 lines to 406 lines.

Tests: 667 (unchanged).
