+++ Sources/BugNarrator/Services/RecordingSessionController.swift
@@ -404,24 +404,3 @@ private enum RecordingTransition {
     case stopping
     case cancelling
 }
-
-enum RecordingSessionStartOutcome {
-    case started(RecordingSessionDraft)
-    case restored(RecordingSessionDraft)
-    case transitionInProgress
-    case busy
-    case preflightFailure(AppError)
-    case failure(Error)
-}
-
-enum RecordingSessionStopReadiness {
-    case ready(RecordingSessionDraft)
-    case transitionInProgress
-    case noActiveRecording
-    case missingSessionMetadata
-}
-
-enum RecordingSessionCancelOutcome {
-    case cancelled(RecordingSessionDraft?)
-    case transitionInProgress
-}
